backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.6
celery==5.4.0
drf-spectacular==0.27.2
uvicorn==0.30.1
gunicorn==22.0.0
# ML
torch==2.3.0
torch-geometric==2.5.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
faiss-cpu==1.8.0